Major polishing
The logout function now works. The login function sort of works as
well, but you can hit the back button to get back to the dashboard so
it’s not even close to secure. Other than that, we seem to be
progressing at a decent pace. Don’t let up guys, we’re close to winning
this war.